Add in playlist support

+++ music_sort.py
@@ -1,17 +1,11 @@
 #! /usr/bin/env python
 
 import mutagen
+import mutagen.mp3
 import os, sys
 import shutil
 import logging
 import re
-
-logging.basicConfig(format='%(asctime)s %(message)s', 
-    datefmt='%m/%d/%Y %H:%M:%S', 
-    filename='/var/log/music_sort.log', level=logging.DEBUG)
-#logging.basicConfig(format='%(asctime)s %(message)s', 
-#    datefmt='%m/%d/%Y %H:%M:%S', 
-#    filename='./log.log', level=logging.DEBUG)
 
 def sanitize (string):
     return re.sub('[/\\\?\*:]', '_', string)
@@ -19,43 +13,17 @@ def sanitize (string):
 UNSORTED_DIR= "/media/raptor/Music-Inbox/"
 SORTED_DIR  = "/media/raptor/Music/"
 ERROR_DIR   = "/media/raptor/Music-Errors/"
-BLANK_TAG   = {'blank': True}
 
 #UNSORTED_DIR= './Unsorted/'
 #SORTED_DIR  = './Sorted/'
 #ERROR_DIR   = './Error/'
 
-if len(os.listdir(UNSORTED_DIR)) == 0:
-    sys.exit()
+BLANK_TAG   = {'blank': True}
 
-logging.info("---")
-logging.info("Starting new music scan...")
-logging.info("")
+def get_tags (unsorted_file, unsorted_dir=UNSORTED_DIR, error_dir=ERROR_DIR):
 
-
-unsorted    = os.listdir(UNSORTED_DIR)
-found_dir   = True
-folders     = []
-while found_dir:
-    found_dir   = False
-    newunsorted = []
-    for entry in unsorted:
-        if os.path.isdir(UNSORTED_DIR + entry):
-            logging.info("Expanding into directory %s", entry)
-            folders.insert(0, entry)
-            newunsorted.extend([entry + "/" + x for x in 
-                    os.listdir(UNSORTED_DIR + entry)])
-            found_dir = True
-        else:
-            newunsorted.append(entry)
-    unsorted = newunsorted[:]
-
-if unsorted:
-    logging.info("Sorting %d files: " + str(unsorted), len(unsorted))
-
-for unsorted_file in unsorted:
     try:
-        tags = mutagen.File(UNSORTED_DIR + unsorted_file, easy=True)
+        tags = mutagen.File(unsorted_dir + unsorted_file, easy=True)
     except mutagen.mp3.HeaderNotFoundError, why:
         tags = BLANK_TAG
         logging.info(" ** Could not get mutagen to read headers: " 
@@ -63,14 +31,18 @@ for unsorted_file in unsorted:
     except IOError, why:
         logging.info(" ** Could not get mutagen to read file: " 
             + unsorted_file + "\n" + str(why))
-        continue
+        return None
 
     if not tags:
-        shutil.move(UNSORTED_DIR + unsorted_file, 
-                ERROR_DIR + os.path.split(unsorted_file)[1])
+        shutil.move(unsorted_dir + unsorted_file, 
+                error_dir + os.path.split(unsorted_file)[1])
         logging.warning(" ** Unable to find music interpreter for file "
              + unsorted_file)
-        continue
+        return None
+
+    return tags
+
+def process_tags (unsorted_file, tags):
 
     if 'albumartist' in tags:
         artist = tags['albumartist'][0]
@@ -99,20 +71,81 @@ for unsorted_file in unsorted:
     album = sanitize(album)
     artist= sanitize(artist)
 
-    logging.info("    " + SORTED_DIR + artist + '/' + album + '/' + title)
+    return (artist, album, title)
 
+# expand recursively into the directories
+def get_all_files_in_dir(directory=UNSORTED_DIR):
+    if directory[-1] != '/':
+        directory = directory + '/'
+    unsorted    = os.listdir(directory)
+    found_dir   = True
+    folders     = []
+    while found_dir:
+        found_dir   = False
+        newunsorted = []
+        for entry in unsorted:
+            if os.path.isdir(directory+ entry):
+                logging.info("Expanding into directory %s", entry)
+                folders.insert(0, entry)
+                newunsorted.extend([entry + "/" + x for x in 
+                        os.listdir(directory+ entry)])
+                found_dir = True
+            else:
+                newunsorted.append(entry)
+        unsorted = newunsorted[:]
+    return (unsorted, folders)
+
+def sort_file(unsorted_file, unsorted_dir=UNSORTED_DIR, 
+        sorted_dir=SORTED_DIR, error_dir=ERROR_DIR):
+
+    tags = get_tags(unsorted_file)
+    if tags is None:
+        return False
+
+    (artist, album, title) = process_tags(unsorted_file, tags)
+
+    logging.info("    " + sorted_dir + artist + '/' + album + '/' + title)
     try:
-        os.makedirs(SORTED_DIR + artist + '/' + album)
+        os.makedirs(sorted_dir + artist + '/' + album)
     except OSError, why:
         logging.info('    OSERROR: ' + str(why))
     
     try:
-        shutil.move(UNSORTED_DIR + unsorted_file, 
-            SORTED_DIR+artist + '/' + album + '/' + title)
+        shutil.move(unsorted_dir + unsorted_file, 
+            sorted_dir+artist + '/' + album + '/' + title)
     except IOError, why:
-        shutil.move(UNSORTED_DIR + unsorted_file, ERROR_DIR + title)
+        shutil.move(unsorted_dir + unsorted_file, error_dir + Title)
         logging.warning(' ** IOERROR: ' +  str(why))
+        return False
+    return True
 
-logging.info("Removing directories " + str(folders))
-for entry in folders:
-    os.rmdir(UNSORTED_DIR + entry)
+
+if __name__ == '__main__':
+
+    #logging.basicConfig(format='%(asctime)s %(message)s', 
+    #    datefmt='%m/%d/%Y %H:%M:%S', 
+    #    filename='/var/log/music_sort.log', level=logging.DEBUG)
+    logging.basicConfig(format='%(asctime)s %(message)s', 
+        datefmt='%m/%d/%Y %H:%M:%S', 
+        filename='./music_sort.log', level=logging.DEBUG)
+
+    if len(os.listdir(UNSORTED_DIR)) == 0:
+        sys.exit()
+
+    logging.info("---")
+    logging.info("Starting new music scan...")
+    logging.info("")
+
+
+    (unsorted, folders) = get_all_files_in_dir()
+
+    if unsorted:
+        logging.info("Sorting %d files: " + str(unsorted), len(unsorted))
+
+    # process files
+    [sort_file(unsorted_file) for unsorted_file in unsorted]
+
+    # post process
+    logging.info("Removing directories " + str(folders))
+    for entry in folders:
+        os.rmdir(UNSORTED_DIR + entry)
